+++ Source/Millistream.Streaming/NativeUnixImplementation.cs
@@ -16,7 +16,7 @@ namespace Millistream.Streaming
             public static extern void mdf_destroy(IntPtr handle);
 
             [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
-            public static extern int mdf_get_next_message(IntPtr handle, ref int message, ref int message_class, ref uint instrument);
+            public static extern int mdf_get_next_message(IntPtr handle, ref int message, ref int message_class, ref ulong instrument);
 
             [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
             public static extern int mdf_get_next_field(IntPtr handle, ref uint tag, ref IntPtr value);
@@ -91,7 +91,7 @@ namespace Millistream.Streaming
         public void mdf_destroy(IntPtr handle) => NativeUnixMethods.mdf_destroy(handle);
         public void mdf_disconnect(IntPtr handle) => NativeUnixMethods.mdf_disconnect(handle);
         public int mdf_get_next_field(IntPtr handle, ref uint tag, ref IntPtr value) => NativeUnixMethods.mdf_get_next_field(handle, ref tag, ref value);
-        public int mdf_get_next_message(IntPtr handle, ref int message, ref int message_class, ref uint instrument) => NativeUnixMethods.mdf_get_next_message(handle, ref message, ref message_class, ref instrument);
+        public int mdf_get_next_message(IntPtr handle, ref int message, ref int message_class, ref ulong instrument) => NativeUnixMethods.mdf_get_next_message(handle, ref message, ref message_class, ref instrument);
         public int mdf_get_property(IntPtr handle, MDF_OPTION option, ref IntPtr value) => NativeUnixMethods.mdf_get_property(handle, option, ref value);
         public int mdf_message_add(IntPtr message, ulong instrument_reference, int message_reference) => NativeUnixMethods.mdf_message_add(message, instrument_reference, message_reference);
         public int mdf_message_add_date(IntPtr message, uint tag, string value) => NativeUnixMethods.mdf_message_add_date(message, tag, value);
